There are three trust anchor ID representations, not two (#40)
Adding on to https://github.com/davidben/tls-trust-expressions/pull/37,
also based on feedback from Russ Housley.

(Whether we actually want three reprs, I dunno. This is perhaps
excessive. We could just decide there are only binary and text reprs and
that, if you stick them in an ASN.1 structure, forget their OID heritage
and put it in an OCTET STRING. It just changes the tag. But, for now,
just describe the thing we got. This is not a particular important
bikeshed.)