Expand on the certificate_authorities discussion in the explainer
The explainer currently says the old one is impractical, which is not
true without qualifications. It's perfectly practical for a large class
of applications. We're just interested in the cases where it is not.

Also, for the many folks who already use certificate_authorities (client
certs), it's worth emphasizing that all these extensions coexist just
fine.