Trim the ACME section a bit
Based on feedback from Aaron Gable, let's just define the media type and
how to use it with the existing ACME mechanisms. In particular, the
alternate chain mechanism is not quite ideal for parallel issuance.
Parallel issuance can be better solved in a separate document, as it's
really orthogonal to the exist trust anchor negotiation strategy.

There's probably other bits we can trim, but I've left them alone for
now.