Slight tweak to wording of X.509 extension (#86)
I think it's better to say the trust anchor is *represented* as an X.509
certificate rather than that it *is* an X.509 certificate. You might
represent the "same" trust anchor in different ways. RFC 5280 doesn't do
a great job of nailing this down.

Also drop the "self-signed" because it might be, e.g.,
draft-davidben-x509-alg-none. (But I don't think it actually makes sense
to cite that draft, so just leave it unspecified.)