Move CertificatePropertyList definition to later (#83)
Playing with a slightly different structure. We actually don't refer to
this structure much. The thinking here is that, like the ACME extension
and the media type, this isn't really fundamental to the extension, just
a RECOMMENDED way to deploy it. This lets us get to the meat of the
mechanism sooner.

I also demoted it from the glossary, since it's now only relevant for
the "optional but recommended" section.