Prepare repository for transition to tlswg
Remove draft-davidben-tls-trust-expr (still in git history if anyone
wants it for reference), and preface most of the supplementary documents
with a historical note.

This does not (yet) rename the document, or update URLs. I figure that's
best done after we transfer + rename the repository. But we can get this
part sorted out first.

Also close out all the old trust-exprs issues:

* Closes #26
* Closes #25
* Closes #17
* Closes #13
* Closes #1